Move jobs update to Jobs class
